파라미터 타입 종류 설명, age를 Long으로 받기

+++ src/main/java/com/ll/exam/sb_batch/job/withParam/WithParamJobConfig.java
@@ -40,7 +40,7 @@ public class WithParamJobConfig {
     @StepScope
     public Tasklet withParamStep1Tasklet(
             @Value("#{jobParameters['name']}") String name,
-            @Value("#{jobParameters['age']}") int age
+            @Value("#{jobParameters['age']}") Long age
     ) {
         return (contribution, chunkContext) -> {
             log.debug("name : {}, age : {}", name, age );
